+++ lib/database_manager.dart
@@ -4,3 +4,5 @@ export 'src/schema/schema.dart';
 export 'src/schema/blueprint.dart';
 export 'src/migration_interface.dart';
 export 'src/migrate.dart';
+
+export 'src/database/sqflite_helper.dart';