+++ CHANGELOG.md
@@ -1,4 +1,7 @@
+## [0.0.2] - 2020-04-18.
+* Created ConnectHelper Class. Possibility to manage the connection with SQflite.
+
 ## [0.0.1] - 2020-04-9.
 * Created Database Manager.
 * Created SQLCommand.
-* Created Defined Schema Class.
+* Created Schema Class.